Fix copy method using mkvmerge
Add a workaround for the problem of mkvmerge working
with transport stream input files having a huge audio offset
in the resulting files.
Now we convert the whole file with ffmpeg into a mkv file
and use mkvmrge on this intermediate file to split and copy
parts into the final video file.